Big refactor and cleanup
Modules have been stripped-out of its services, since we are not doing
something even similar to a Service Oriented Architecture, and just splitting
them as mere Java libraries works as intended. This implies that only one
service publishes an HTTP API, that is now named "mahmi-http".

Database access module (mahmi-database) has been all refactored to use a new
fully independent component "FunctionalJDBC", that allows us to operate with
all those utility functions over JDBC without requiring inheritance.
MySQLAbstractDAO, this way, just does what it should: provide a template to
create DAOs for MySQL databases. MetaGenomesDAO does not longer return a full
metagenome (including the stored Fasta file), since it won't be required in a
nearby future and just consumes too much memory to be done always.

Most of the entities have been replaced from immutable versions to mutable
ones (with setters), because efficiency reasons and because Java has not been
designed with immutability in mind, and some operations become quite cumbersome
when done in a purely functional manner.

Protein cuttter module (mahmi-cutter) now just receives the Project entity of
which recover all associated Proteins to database, and digest them, instead of
receiving a complete set of Proteins.

A lot of smaller changes have been made all over the project, this is a big
commit, and implies a big design change. Future commits will no longer be this
big, and working with a flow of small changes will be a requirement.