Move DNA validity check out of constructor
DNASequence monoid has performance problems when we need to recheck every time
(in the DNASequence constructor) a sum is performed whether or not a sequence
is valid. Since we are just appending already constructed sequences, we are
sure that those previously created are valid, and thus can remove that extra
checking after concatenating them.

This has been done by just performing the check in the public smart
constructors (static), that are the only ones allowed to create a new
DNASequence object. This way, since the Monoid uses the private constructor
with the concatenated sequence, we are not rechecking it again.