Fix CutProteinsWrapper unmarshalling of enzymes
Requests to HTTP API have been found unresolved whenever a "digest proteins"
order was sent. The bug was caused by incorrect JAXB marshalling/unmarshalling
of the "enzymes" field in the object "CutProteinsWrapper".

It looks like JAXB is unable to unmarshall a set of enzymes when typed as a Set
interface (Set<Enzyme>) and also cannot parse it as expected (with an upper
"<enzymes></enzymes>" xml element, and each enzyme being its own xml element
"<enzyme></enzyme>" inside it) without @XmlElementWrapper and @XmlElements
annotations. It has been solved by using a LinkedHashSet implementation instead
of the interface and annotating the field with that mentioned annotations.