Make Fasta and FastaReader pseudo-lazy
Fasta is now a Iterable of CompoundSequences, with no memoization. That means
that reading the next element will discard the previous one, thus only
maintaining one sequence of the Fasta file in memory. Fasta also has a
toStream() method that should be called if a memoized collection of sequences
is required.

FastaReader and FastaWriter do no longer support the "fromFile" static
constructor, and a "fromPath" should always be used instead.

FunctionalJDBC::fasta has also been changed to just write a String (created
through StringWriter) instead of piping an OutputStream to a InputStream. It
seems more convenient.