Upgrade lombok and reorganize imports
Lombok has been upgraded from 1.16.0 to 1.16.2, and some of the maven errors
have also been manually fixed (e.g.: by not using val) [see #6].

Imports of classes have been reorganized to be more meaningful and correct,
this is the order now:

    import java.*
    import javax.*
    import lombok.*
    import fj.*
    import org.*
    import com.*
    import es.uvigo.ei.sing.*
    import static java.*
    import static javax.*
    import static fj.*
    import static es.uvigo.ei.sing.*
    import static *

So, basically: functionaljava, lombok and our own sing imports are now
correctly sorted. Moreover, all static imports have been moved below the
non-static ones. This is still not a style guideline, but will probably be the
recommended way to go in a nearby future.