Fix SHA1.{hash,equal}, make CompoundSequence use them
Previous commit fixed CompoundSequence hash and equal checks by not using the
broken ones in SHA1 but by directly using the SeqHash and SeqEqual and
comparing each Compound. This is a much more slower method than the previous
one.

SHA1 has been fixed to use StringEqual instead of ArrayEqual<ByteEqual>, since
something in it seems broken. Now the comparison is done through the
hexadecimal representation of the SHA1 digest instead of the inner byte array.
And with that, CompoundSequence has been reverted to use SHA1 hash and equal
instead of the complete sequence equality check. This makes CompoundSequence
hash/equal checks to be as efficient as the previous SHA1 with bytearray ones.